backend/pyproject.toml
[tool.poetry]
name = "litesearch"
version = "0.1.0"
description = ""
authors = ["meirdev <maornet@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.40"}
aiosqlite = "^0.17.0"
greenlet = "^1.1.2"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
lark = "^1.1.2"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
mypy = "^0.971"
SQLAlchemy = {extras = ["mypy"], version = "^1.4.40"}
flake8 = "^5.0.4"
requests = "^2.28.1"
Faker = "^14.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
allow_redefinition = true
plugins = "sqlalchemy.ext.mypy.plugin"

[tool.isort]
profile = "black"
